feature(core) fixes after CR
