+++ module/core/src/Application/Exception/NormalizationException.php
@@ -7,12 +7,12 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Core\Infrastructure\Exception;
+namespace Ergonode\Core\Application\Exception;
 
-class NormalizerException extends \Exception
+class NormalizationException extends NormalizerException
 {
     public function __construct(string $message, \Throwable $previous = null)
     {
-        parent::__construct($message, 0, $previous);
+        parent::__construct($message, $previous);
     }
 }
