+++ module/core/src/Application/Serializer/JMSSerializer.php
@@ -8,21 +8,26 @@ declare(strict_types=1);
 
 namespace Ergonode\Core\Application\Serializer;
 
-use JMS\Serializer\SerializerInterface as JMSSerializerAlias;
+use Ergonode\Core\Application\Serializer\NormalizerInterface;
+use Ergonode\Core\Infrastructure\Exception\DenoralizationException;
+use Ergonode\Core\Infrastructure\Exception\NormalizerException;
+use JMS\Serializer\Serializer;
 use Psr\Log\LoggerInterface;
 use Ergonode\Core\Infrastructure\Exception\SerializationException;
 use Ergonode\Core\Infrastructure\Exception\DeserializationException;
 
-class JMSSerializer implements SerializerInterface
+class JMSSerializer implements SerializerInterface, NormalizerInterface
 {
     private const SERIALIZE = 'Can\'t serialize object "%s" to "%s" format';
     private const DESERIALIZE = 'Can\'t deserialize data "%s" as "%s" from "%s" format';
+    private const NORMALIZE = 'Can\'t normalize object "%s"';
+    private const DENORMALIZE = 'Can\'t denormalize data from "%s" to "%s" type';
 
-    private JMSSerializerAlias $serializer;
+    private Serializer $serializer;
 
     private LoggerInterface $logger;
 
-    public function __construct(JMSSerializerAlias $serializer, LoggerInterface $logger)
+    public function __construct(Serializer $serializer, LoggerInterface $logger)
     {
         $this->serializer = $serializer;
         $this->logger = $logger;
@@ -57,6 +62,34 @@ class JMSSerializer implements SerializerInterface
 
             throw new DeserializationException(
                 sprintf(self::DESERIALIZE, $data, $type, $format),
+                $exception
+            );
+        }
+    }
+
+    public function normalize(object $data, ?string $type): array
+    {
+        try {
+            return $this->serializer->toArray($data, null, $type);
+        } catch (\Throwable $exception) {
+            $this->logger->error($exception);
+
+            throw new NormalizerException(
+                sprintf(self::NORMALIZE, get_debug_type($data)),
+                $exception
+            );
+        }
+    }
+
+    public function denormalize(array $data, string $type): object
+    {
+        try {
+            return $this->serializer->fromArray($data, $type);
+        } catch (\Throwable $exception) {
+            $this->logger->error($exception);
+
+            throw new DenoralizationException(
+                sprintf(self::DENORMALIZE, get_debug_type($data), $type),
                 $exception
             );
         }
